% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyderogen.R
\docType{package}
\name{hyderogen-package}
\alias{hyderogen}
\alias{hyderogen-package}
\title{Build jekyll sites}
\description{
hyderogen provides a toolset for quickly building Jekyll projects from your
package's roxygen. These tools focus on creating detailed front matter and
plain templates. You can then further customize these templates. When you
update your roxygen you can rebuild your Jekyll front matter, while leaving
the templates untouched.

This package is a work in progress. If you try it out please do not be shy
about reporting bugs or less-than-ideal results. All feedback is welcome at
\href{https://github.com/nteetor/hyderogen}{nteetor/hyderogen/issues}.
}
\seealso{
For more on Jekyll check out \url{https://jekyllrb.com/docs/home/}.

For more on using Jekyll and Github pages check out,
\url{https://jekyllrb.com/docs/github-pages/}.
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
