% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyderogen.R
\docType{package}
\name{hyderogen}
\alias{hyderogen}
\alias{hyderogen-package}
\title{hyderogen: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
