% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jekyll.R
\name{jekyll}
\alias{jekyll}
\title{Jekyll folder structure}
\usage{
jekyll(path = ".", dir = "docs", build = FALSE)
}
\arguments{
\item{path}{A file path specifying the folder of a package, defaults to
\code{"."}.}

\item{dir}{A file path, relative to \code{path}, specifying the destination of the
jekyll folders and files, defaults to \code{"docs/"}.}

\item{build}{One of \code{TRUE} or \code{FALSE} specifying if the jekyll site is built
after creating the folder structure, defaults to \code{FALSE}.}
}
\description{
Create a jekyll site structure from a package's roxygen. Optionally, a
site may be built, too. Building your site requires
\href{https://jekyllrb.com/docs/installation/}{jekyll} to be installed and
available on your system.
}
