% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jekyll.R
\name{jekyll}
\alias{jekyll}
\title{Build a jekyll site}
\usage{
jekyll(pkg = ".", dir = "docs")
}
\arguments{
\item{pkg}{A character string specifying a file path to an R package,
defaults to \code{"."}, the current directory.}

\item{dir}{A folder path, relative to \code{pkg}, where the jekyll site will be
built, defaults to \code{"docs"}.}
}
\description{
The \code{jekyll} function builds a jekyll site from your
package's roxygen.
}
\details{
Collections are currently overwritten. However, if a family's name is
changed then the old, corresponding collection will not be deleted. This is
to prevent custom collections from accidentally being removed. This behaviour
is subject to change.
}
\examples{
\dontrun{

jekyll("~/git/mypkg")

}

}
